Extract BccUploadMetricsTests
